# openapi.yaml
openapi: 3.0.0
info:
  title: Service Envelope API v2.0
  version: 2.1.0
  description: REST API for the Service Envelope System. Includes OTP bearer
    authentication, Mongo-backed auth users and roles, generic email template
    fallbacks, schema v1.0.5 service validation, and unified delivery selection.
  contact:
    name: Service Envelope Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development
  - url: https://api.mapua.edu.ph
    description: Production
paths:
...
  /api/auth/verify:
    post:
      tags:
        - Auth
      summary: Verify Bearer Token
      operationId: verifyToken
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  user:
                    type: object
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    valid: true
                    user:
                      email: barondimaranan@gmail.com
                      role: super_admin
                      name: Baron Dimaranan
                    expiresIn: 1h
        "401":
          $ref: "#/components/responses/Unauthorized"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: 'curl -X POST "http://localhost:8000/api/auth/verify" -H "Authorization: Bearer $TOKEN"'
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get Current Bearer User
      operationId: getCurrentAuthUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUserPublic"
                  expiresIn:
                    type: string
                    example: 1h
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/OTP/flush:
    post:
      tags:
        - Auth
      summary: Flush Stale OTP Challenges
      operationId: flushOtp
      security:
        - BearerAuth: []
      description: Requires a super_admin bearer token.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                consumedRetentionHours:
                  type: number
                  default: 24
      responses:
        "200":
          description: Stale OTP records flushed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpFlushResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/services:
    get:
      tags:
        - Services
      summary: List All Services
      operationId: listAllServices
      security: []
      responses:
        "200":
          description: All available services
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 4
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceSummary"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    count: 4
                    services:
                      - serviceId: SERVICE-999
                        id: SERV-1-04152026
                        name: sample
                        description: sample
                        type: sample
                        yaml: sample
        "500":
          $ref: "#/components/responses/ServerError"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET "http://localhost:8000/api/services"
...
  /api/requests:
    get:
      tags:
        - Requests
      summary: List All Requests
      operationId: listAllRequests
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusQueryParam"
        - $ref: "#/components/parameters/TypeQueryParam"
        - $ref: "#/components/parameters/LimitQueryParam"
        - $ref: "#/components/parameters/OffsetQueryParam"
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    total: 1
                    count: 1
                    limit: 1
                    offset: 1
                    requests:
                      - id: sample
                        type: sample
                        initiator: sample
                        status: sample
                        createdAt: 2026-06-07T00:00:00.000Z
                        lastUpdated: 2026-06-07T00:00:00.000Z
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET "http://localhost:8000/api/requests"
    post:
      tags:
        - Requests
      summary: Create Request
      operationId: createRequest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCreatePayload"
            examples:
              usingServiceId:
                value:
                  serviceId: SERVICE-999
                  initiator: requester@mapua.edu.ph
                  parameters:
                    studentId: 20250001
                    firstName: Juan
                    lastName: Dela Cruz
                    email: juan.delacruz@mapua.edu.ph
              usingType:
                value:
                  type: comprehensive-student-document
                  initiator: requester@mapua.edu.ph
                  parameters:
                    studentId: 20250001
                    firstName: Juan
                    lastName: Dela Cruz
                    email: juan.delacruz@mapua.edu.ph
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRequestResponse"
              examples:
                sample:
                  summary: Sample 201 response
                  value:
                    success: true
                    requestId: REQ-20260606-042
                    type: sample
                    serviceId: sample
                    status: queued
                    envelopes:
                      sampleKey: sampleValue
                    message: sample
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X POST "http://localhost:8000/api/requests" \
              -H "Content-Type: application/json" \
              -d '{"sample":true}'
...
  /api/admin/services:
    get:
      tags:
        - Admin
      summary: Get All Services (Admin)
      operationId: getAllServicesAdmin
      security: []
      responses:
        "200":
          description: All service definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceDefinitionDetails"
                  total:
                    type: integer
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    services:
                      - serviceId: sample
                        id: sample
                        name: sample
                        service_id: sample
                        service_name: sample
                        type: sample
                        description: sample
                        envelopes:
                          sampleKey: sampleValue
                        requestParameters:
                          - sampleKey: sampleValue
                    total: 1
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET "http://localhost:8000/api/admin/services"
    post:
      tags:
        - Admin
      summary: Create Service (Admin)
      operationId: createServiceAdmin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServicePayload"
            examples:
              sample:
                summary: Sample request
                value:
                  name: Transcript Request
                  yaml: sample
                  type: request-tor
                  initiator: student
      responses:
        "201":
          description: Service created
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: >-
            curl -X POST "http://localhost:8000/api/admin/services" \
              -H "Content-Type: application/json" \
              -d '{"name":"Transcript Request","yaml":"sample","type":"request-tor","initiator":"student"}'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StatusQueryParam:
      name: status
      in: query
      required: false
      schema:
        type: string
      example: pending_payment
      description: Filter by request overall status.
    TypeQueryParam:
      name: type
      in: query
      required: false
      schema:
        type: string
      example: comprehensive-student-document
      description: Filter by request type.
    LimitQueryParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 1000
      example: 20
    OffsetQueryParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
...
  schemas:
    ApiRole:
      type: string
      enum:
        - super_admin
        - admin
        - requester
        - orchestrator
      example: requester
    AuthUserPublic:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ApiRole"
        name:
          type: string
...
    OtpFlushResponse:
      type: object
      properties:
        status:
          type: string
          example: flushed
        deletedCount:
          type: integer
    ServiceSummary:
      type: object
      properties:
        serviceId:
          type: string
          example: SERVICE-999
        id:
          type: string
          example: SERV-1-04152026
        name:
          type: string
        description:
          type: string
        type:
          type: string
        yaml:
          type: string
    ServiceDefinitionDetails:
      type: object
      properties:
        serviceId:
          type: string
        id:
          type: string
        name:
          type: string
        service_id:
          type: string
        service_name:
          type: string
        type:
          type: string
        description:
          type: string
        envelopes:
          type: object
          additionalProperties: true
        requestParameters:
          type: array
          items:
            type: object
            additionalProperties: true
    RequestCreatePayload:
      type: object
      required:
        - initiator
        - parameters
      properties:
        type:
          type: string
          description: Service type identifier.
          example: comprehensive-student-document
        serviceId:
          type: string
          description: Service ID identifier.
          example: SERVICE-999
        initiator:
          type: string
          description: Actor creating the request.
          example: requester@mapua.edu.ph
        parameters:
          type: object
          additionalProperties: true
      anyOf:
        - required:
            - type
        - required:
            - serviceId
    CreateRequestResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          example: REQ-20260606-042
        type:
          type: string
        serviceId:
          type: string
        status:
          type: string
          example: queued
        envelopes:
          type: object
          additionalProperties: true
        message:
          type: string
    RequestListResponse:
      type: object
      properties:
        total:
          type: integer
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        requests:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              initiator:
                type: string
              status:
                type: string
              createdAt:
                type: string
                format: date-time
              lastUpdated:
                type: string
                format: date-time
...
    CreateServicePayload:
      type: object
      required:
        - name
        - yaml
        - type
        - initiator
      properties:
        name:
          type: string
          example: Transcript Request
        yaml:
          type: string
          description: Full service definition YAML text.
        type:
          type: string
          example: request-tor
        initiator:
          type: string
          example: student
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Request not found
        details:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 400 response
              value:
                error: Request not found
                details: sample
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 401 response
              value:
                error: Request not found
                details: sample
    Forbidden:
      description: Forbidden - Authenticated user does not have the required role or ownership
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficientRole:
              value:
                error: Insufficient permissions
...
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 500 response
              value:
                error: Request not found
                details: sample
...